+++ src/PoshServerLess/classes/AzFunctionsBinding.class.ps1
@@ -1,0 +1,8 @@
+class AzFunctionsBinding {
+
+    [string] $BindingName
+    [string] $BindingDirection
+    [string] $BindingType
+    
+
+}