+++ src/PoshServerLess/classes/TestModulePresent.ps1
@@ -1,0 +1,5 @@
+
+function TestModulePresent ([string] $moduleName="Az") {
+
+    return ! $null -eq (get-module -ListAvailable | where-object name -eq $moduleName)
+}