+++ src/PoshServerLess/classes/getMsTimeZone.ps1
@@ -1,0 +1,16 @@
+function getMsTimeZone () {
+
+    $ModulePath = $PSScriptRoot 
+
+    try {
+        $jsonArmTemplatePath = Join-Path -Path $ModulePath -ChildPath "function.json"
+        $ObjectData = (Get-Content -Path $jsonArmTemplatePath  -Raw | ConvertFrom-Json)
+
+        return $ObjectData.parameters.timezone.allowedValues
+
+    }
+    catch {
+        Write-Error -Message " Exception Type: $($_.Exception.GetType().FullName) $($_.Exception.Message)"
+    }
+
+}