+++ src/PoshServerLess/classes/queue.class.ps1
@@ -1,0 +1,12 @@
+class queue : AzFunctionsBinding {
+    [string] $queueName
+    [string] $connection
+
+    queue ([string] $Name, [string] $QueuName, [string] $connection) {
+        $this.BindingName = $Name
+        $this.connection = $Connection
+        $this.queueName = $QueuName
+        $this.BindingDirection = "out"
+        $this.BindingType = "queue"
+    }
+}