+++ changelog.md
@@ -1,5 +1,11 @@
 # Release Note
 
+## Unreleased
+
+* Split class in multiple files
+
+## Releases
+
 * V 0.1.4 Add a cmdlet to set App Settings and Timezone
 * V 0.1.4 Add a cmdlet to get App Settings and to load App Setting
 * V 0.1.3 Add a method to extract Function App settings from Azure